pyproject.toml
[tool.poetry]
name = "repl_nix_Video-Game-Dev"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^9.3.0"
numpy = "^1.23.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
